11:59 TUE 14 September

+++ lib/Models/bookingListModal.dart
@@ -278,7 +278,7 @@ class FiestaDetail {
   String? ticketPriceStandard;
   String? ticketPriceVip;
   String? totalMembers;
-  int? clubRating;
+  double? clubRating;
   var leftNormalTicket;
   var leftStandardTicket;
   var leftVipTicket;
